refactor to remove unused and new folders
