+++ lore-search-app/app/ui/bookmarks/bookmarkListItem.tsx
