--- lore-search-app/app/ui/bookmarksTab.tsx
@@ -1,5 +1,0 @@
-
-
-export default function BookmarksTab() {
-    return (<></>);
-}
