+++ lore-search-app/app/ui/search/searchBox.tsx
