+++ lore-search-app/app/ui/search/searchNavigationList.tsx
