+++ lore-search-app/app/ui/search/searchResults.tsx
@@ -1,4 +1,4 @@
-import { fetchSearchResults, fetchSynonyms } from "../lib/data";
+import { fetchSearchResults } from "../../lib/data";
 import LoreEntryList from "./loreEntryList";
 
 export default async function SearchResults({
@@ -12,13 +12,7 @@ export default async function SearchResults({
 }) {
   const results = await fetchSearchResults(query, currentPage, sort);
 
-  const options = [
-    "Try harder next time.",
-    "Maybe something less obscure?",
-    "Check for typos!",
-  ];
-
   return (
-    <LoreEntryList items={results} showBookmark={true} />
+    <LoreEntryList items={results} />
   );
 }
